+++ src/Model/Produto.php
@@ -54,7 +54,8 @@ class Produto
                 'insert',
                 $antes,
                 $depois,
-                $usuario
+                $usuario,
+                $posto
             );
 
             return ['status' => 'success', 'message' => 'Produto cadastrado com sucesso!'];
@@ -63,13 +64,12 @@ class Produto
         return ['status' => 'error', 'message' => 'Erro ao cadastrar produto!'];
     }
 
-    public static function buscarPorCodigo($codigo, $posto)
+    public static function buscarPorProduto($produto, $posto)
     {
         $con = Db::getConnection();
-        $codigo = pg_escape_string($codigo);
         $posto = intval($posto);
 
-        $sql = "SELECT produto, codigo, descricao, ativo FROM tbl_produto WHERE codigo = '{$codigo}' AND posto = {$posto}";
+        $sql = "SELECT produto, codigo, descricao, ativo FROM tbl_produto WHERE produto = {$produto} AND posto = {$posto}";
         $res = pg_query($con, $sql);
 
         return pg_num_rows($res) > 0
@@ -127,7 +127,8 @@ class Produto
                 'update',
                 $antes,
                 $depois,
-                $usuario
+                $usuario,
+                $posto
             );
 
         return ['status' => 'success', 'message' => 'Produto atualizado com sucesso!'];
@@ -157,12 +158,16 @@ class Produto
         $con = Db::getConnection();
         $posto = intval($posto);
 
+        $produto_tem_estoque = self::produtoTemEstoque($produto, $posto);
+
+        if ($produto_tem_estoque == false) {
         $sql = "DELETE FROM tbl_produto WHERE produto = $produto AND posto = $posto";
         $res = pg_query($con, $sql);
 
-        return $res
-            ? ['status' => 'success', 'message' => 'Produto excluido com sucesso']
-            : ['status' => 'error', 'message' => 'Erro ao exlcuir produto.'];
+        return $res ? ['status' => 'success', 'message' => 'Produto excluído com sucesso.'] : ['status' => 'error', 'message' => 'Erro ao excluir produto.'];
+        } else {
+            return ['status' => 'error', 'message' => 'Não é possível excluir. O produto ainda está vinculada ao estoque.'];
+        }
     }
 
     public static function autocompleteProdutos($termo, $posto)
@@ -171,15 +176,15 @@ class Produto
         $termo = pg_escape_string($termo);
         $posto = intval($posto);
 
-        // busca por descrição ou código (somente do posto); filtra ativos se quiser
-        $sql = "
-            SELECT produto, codigo, descricao
-              FROM tbl_produto
-             WHERE posto = {$posto}
-               AND (descricao ILIKE '%{$termo}%' OR codigo ILIKE '%{$termo}%')
-             ORDER BY descricao
-             LIMIT 20
-        ";
+        $sql = "SELECT produto,
+                       codigo,
+                       descricao
+                FROM tbl_produto
+                WHERE posto = {$posto}
+                AND (descricao ILIKE '%{$termo}%' OR codigo ILIKE '%{$termo}%')
+                ORDER BY descricao
+                LIMIT 20
+            ";
 
         $res = pg_query($con, $sql);
         $sugestoes = [];
@@ -193,5 +198,26 @@ class Produto
             ];
         }
         return $sugestoes;
+    }
+
+    private static function produtoTemEstoque($produto, $posto)
+    {
+        $con = Db::getConnection();
+        $produto = intval($produto);
+        $posto = intval($posto);
+
+        $sql = "SELECT qtde
+                FROM tbl_estoque
+                WHERE produto = $produto
+                AND posto = $posto
+            ";
+        $res = pg_query($con, $sql);
+
+        if (pg_num_rows($res) > 0) {
+            $row = pg_fetch_assoc($res);
+            return $row['qtde'] > 0;
+        }
+
+        return false;
     }
 }
